fixing home page login bug
