+++ src/types/LoginForm.ts
@@ -1,3 +1,3 @@
 import { loginSchema } from "@/schemas/loginSchema";
 import z from "zod";
-export type LoginForm = z.infer<typeof loginSchema>;
+export type LoginFormType = z.infer<typeof loginSchema>;
